requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
openai==1.3.7
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.1.3
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1
celery==5.3.4
redis==5.0.1
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-wordpress-xmlrpc==2.3
markdown==3.5.1
pillow==10.1.0
selenium==4.15.2
webdriver-manager==4.0.1
lxml==4.9.3
fake-useragent==1.4.0